# docs/swagger/facebook-ads-swagger.yaml
openapi: 3.0.3
info:
  title: Marketing Hub - Facebook Ads API
  version: 1.0.0
  description: |
    Contrato dos endpoints do backend que atendem a integração com Facebook Ads.
    A implementação canônica dos controllers fica concentrada no pacote
    `com.marketinghub.facebookads.controller`, mantendo as rotas públicas legadas
    para compatibilidade com frontend e workers.
...
paths:
...
  /api/accounts/facebook/{id}:
    put:
      tags: [Facebook Accounts]
      summary: Atualiza uma conta do Facebook sem descartar campos não enviados pela UI.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Conta atualizada. } }
    delete:
      tags: [Facebook Accounts]
      summary: Remove uma conta do Facebook.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '204': { description: Conta removida. } }
...
  /api/accounts/facebook/{id}/token/renewal:
    post:
      tags: [Facebook Accounts]
      summary: Registra resultado de renovação do token da conta.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Renovação registrada. } }
  /api/accounts/facebook/{id}/token/revalidation:
    post:
      tags: [Facebook Accounts]
      summary: Revalida o token da conta com a integração Meta.
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { '200': { description: Revalidação registrada. } }
...